Correcting synchronous dynamics event generation
In the previous implementation, the calls were being made against `proc` which is the Process which doesn't implement that function. It seems to need to be accessed through the dynamics on this process which resolves an Attribute Error which prevent tests from running. Additionally, the function was previously passed `t` which is the current time step of the simulation but as events are generated from the dynamics on a per-process basis, it needs to be passed the process instead to retrieve the correct set of results.
83 unit tests pass